Decrease the rate for sampled dedupe logging.
We've gained good confidence that deduplication actually
works correctly, and the dedupe rate has grown significantly
as we've been switching jobs, so it started polluting the logs.
decrease the rate of sampled logging for dedupe 10 times,
we still want to have some sampled events in the logs for sanity
checking.

Change-Id: I19e9f80eb17e6fa8f2df12a5ff07a2c259a26b4d