Switch LocalRenameUserJob for all wikis.
The LocalRenameUserJob schedules a new job for the next
wiki in the list after it's done executing for the current
wiki. With this mechanism the job for meta.wiki is scheduled in
the context of mediawiki, so it only ends up in kafka, while
it's not executed with the kafka queue because it is not enabled.

So our partial wiki-by-wiki transition method doesn't work for
the LocalRenameUserJob, so it needs to be switched globally.

Bug: T193254
Change-Id: I38a1db6f744fe13bbd7869d8c3faecf9e10ea8cf