Enable cirrus jobs for all but wikipedia, commons and wikidata.
Cirrus doesn't use cross-wiki posting, so it's safe to enable
jobs for a subset of wikis.

For all the jobs except ElasticaWrite and a checker job the transition
should be trivial, as they do not use any non-standard queue features.

For ElasticaWrite we will still rely on the retry mechanism used in
the old queue. After enabling compression, the volume should not be
an issue, and if it becomes an issue, we will increase the retry delay.
Also, change-prop retries are switched off for cirrus as it manages it's
retries itself. The problem with pausing the processing will be mitigated
with increasing the retry delays and relying on limited concurrency.

For checker job we set the `reenqueue_delay` to one hour, so each checker
job might be reenqueued only once. WE will need to play with concurrencies
for these jobs, but eventually we will be able to achieve exactly the same
effect as we have right now in cirrus for these jobs by using concurrencies.

Bug: T189137
Change-Id: I494345ac0e239d5703d2499a28ba5209fc9c5381