Decrease the checkerJob delays recheck to 10 minutes.
The search sanitizer enqueues the checkerJobs with random delays
within the checker period. So, with an 1 hour reenqueue delay we
basically can only process that many jobs every 1 hour until
the concurrency limit is not filled up with the jobs from the
future. This creates an artificially large backlog for these
jobs and makes them bursty.

A more proper solution would be to sort the jobs before posting a
bulk of them, then we can increser the reenqueue delay again, but
we first need to deal with 30k jobs accumulated in the queue.

Change-Id: Ie3334d298280deb43241b71c7da3ac15992db742