Fine-tune the concurrencies for jobs.
If the default concurrency used for the low-traffic jobs
is too high, the jobs are running way below the concurrency
limit. When execution becomes slow, they have a huge room
to pile up, using more and more DB connections and HHVM threads.

To avoid that in the short term, we need to better fine-tune the
concurrencies.

We now have the numbers of real used concurrencies within the system,
so we can fine-tune the settings appropriately.

Bug: T202107
Change-Id: I70a7ac032c5a37bb37e081b51038bf0bf349f706